tests/requirements.txt
# Testing requirements
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
pytest-mock==3.12.0
pytest-xdist==3.3.1
pytest-timeout==2.2.0
pytest-benchmark==4.0.0
pytest-html==4.1.1

# Test data and factories
factory-boy==3.3.0
Faker==19.13.0

# HTTP and API testing
httpx==0.25.2
requests-mock==1.11.0

# Database testing
pytest-postgresql==5.0.0
pytest-redis==3.0.2

# Async testing
aioresponses==0.7.4
asynctest==0.13.0

# Load testing
locust==2.17.0
artillery==1.0.0

# Mocking and fixtures
responses==0.24.1
freezegun==1.2.2
pytest-freezer==0.4.8

# Performance testing
memory-profiler==0.61.0
py-spy==0.3.14

# Security testing
bandit[toml]==1.7.5
safety==2.3.5

# Test utilities
testcontainers[postgresql,redis]==3.7.1
moto[s3]==4.2.14
fakeredis==2.20.1

# Coverage and reporting
coverage[toml]==7.3.2
pytest-json-report==1.5.0